sleeper-companion-app/requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
Flask==3.1.2
greenlet==3.2.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
playwright==1.54.0
pyee==13.0.0
python-dotenv==1.1.1
requests==2.32.5
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
